favicon, move carousel to shared, make footer smaller, update logo
