+++ public/index.html
@@ -3,6 +3,7 @@
     <head>
         <meta charset="UTF-8" />
         <title>Playnet</title>
+        <link rel="shortcut icon" href="favicon.svg"/>
     </head>
     <body>
         <div id="root"></div>
