+++ src/components/Home/index.tsx
@@ -1,13 +1,12 @@
 import React from 'react';
 
-import Carousel from '../Carousel';
-
 import Intro10Data from '../LottieAnimations/Intro10Animation';
 import Intro11Data from '../LottieAnimations/Intro11Animation';
 import Intro12Data from '../LottieAnimations/Intro12Animation';
 import Intro8Data from '../LottieAnimations/Intro8Animation';
 import Intro9Data from '../LottieAnimations/Intro9Animation';
 import Base from '../shared/Base';
+import Carousel from '../shared/Carousel';
 import LottieControl from '../shared/LottieControl';
 import { HeaderSections } from '../shared/PlaynetConstants';
