+++ src/components/shared/Carousel.tsx
@@ -1,8 +1,9 @@
 import React, { useEffect, useState } from 'react';
 
-import './styles/Carousel.scss';
-import NextSvg from '../assets/next_btn.svg';
-import PrevSvg from '../assets/prev_btn.svg';
+import '../styles/Carousel.scss';
+
+import NextSvg from '../../assets/next_btn.svg';
+import PrevSvg from '../../assets/prev_btn.svg';
 
 export const CarouselContext = React.createContext({
   next: (): void => undefined,
