+++ src/components/styles/Footer.scss
@@ -5,7 +5,7 @@
   bottom: 0;
   z-index: 99;
   width: 100%;
-  height: $head-foot-height;
+  height: $foot-height;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -27,7 +27,7 @@
   left: 0;
   bottom: 0;
   font: $font-normal;
-  font-size: 12px;
+  font-size: 10px;
   padding: 0;
   margin: 5px 10px;
 }
