+++ src/components/styles/_variables.scss
@@ -9,5 +9,6 @@ $font-title:  normal 700 42px 'Nunito', sans-serif;
 $bg-footer: #fff;
 
 $head-foot-height: 64px;
+$foot-height: 55px;
 
 $playnet-orange: #ff5c41;
