+++ webpack.common.js
@@ -31,6 +31,7 @@ module.exports = {
   plugins: [
     new HtmlWebpackPlugin({
       template: './public/index.html',
+      favicon: './public/favicon.svg',
       filename: './index.html',
     }),
     new CleanWebpackPlugin(),
